Updating objects

+++ models/storage_partition.go
@@ -3,9 +3,9 @@ package models
 type StoragePartition struct {
 	Alias             string
 	Name              string
-	MaxSize           int
+	MaxSize           int64
 	MaxObjects        int
-	CurrentSize       int
+	CurrentSize       int64
 	CurrentObjects    int
 	Id                string
 	StorageLocationId string
